backend/requirements.txt
fastapi
uvicorn
supabase
intuit-oauth
python-quickbooks
cryptography
python-dotenv
pydantic
httpx
pydantic-settings
